Added code for computing thread-avg thruput.
Two new functions in utils.timing. The first one computes the event loop
time averaged over threads, and the other computes the throughput using
this new event loop time calculation. This throughput calculation should
be more stable for fluctuations due to idle cores at the end of the
event loop, since the idle time is ignored in this case.